+++ src/vexpress_boot.py
@@ -25,8 +25,9 @@ def make():
         return
 
     if not os.path.isdir(linux_folder):
-        pretty_print.step("download linux")
-        urllib.request.urlretrieve(url, filename=linux_zip, reporthook=pretty_print.ReportHook().reporthook)
+        if not os.path.isfile(linux_zip):
+            pretty_print.step("download linux")
+            urllib.request.urlretrieve(url, filename=linux_zip, reporthook=pretty_print.ReportHook().reporthook)
 
         pretty_print.step("extract linux")
         zipFile = ZipFile(linux_zip)
